+++ tests/testthat/test_mutual_within.R
@@ -16,7 +16,7 @@ test_that("dimensions and bootstrapping", {
 
     within_se <- mutual_within(test_data, "u", "g",
                                within = "supergroup", weight = "n", se = TRUE)
-    expect_equal(dim(within_se), c(2 * 4, 4))
+    expect_equal(dim(within_se), c(2 * 4, 5))
 })
 
 test_that("bootstrapping fails when sample size is non-integer", {
@@ -34,7 +34,7 @@ test_that("bootstrapping fails when sample size is non-integer", {
     test_data$n2 <- test_data$n / sum(test_data$n) * round(sum(test_data$n))
     ret <- mutual_within(test_data, "u", "g",
             within = "supergroup", weight = "n2", se = TRUE)
-    expect_equal(dim(ret), c(2 * 4, 4))
+    expect_equal(dim(ret), c(2 * 4, 5))
 })
 
 test_that("between + within = total", {
@@ -62,17 +62,17 @@ test_that("option wide works", {
     nowide <- mutual_within(test_data, "u", "g",
                             within = "supergroup", weight = "n")
     nowide_se <- mutual_within(test_data, "u", "g",
-                            within = "supergroup", weight = "n", se = T)
+                            within = "supergroup", weight = "n", se = TRUE)
     wide <- mutual_within(test_data, "u", "g",
-                            within = "supergroup", weight = "n", wide = T)
+                            within = "supergroup", weight = "n", wide = TRUE)
     wide_se <- mutual_within(test_data, "u", "g",
-                            within = "supergroup", weight = "n", wide = T, se = T)
+                            within = "supergroup", weight = "n", wide = TRUE, se = TRUE)
 
-    expect_equal(ncol(nowide) + 1, ncol(nowide_se))
+    expect_equal(ncol(nowide) + 2, ncol(nowide_se))
     expect_equal(nrow(nowide), 2 * 4)
     expect_equal(nrow(nowide), nrow(nowide_se))
 
-    expect_equal(ncol(wide) + 4, ncol(wide_se))
+    expect_equal(ncol(wide) + 8, ncol(wide_se))
     expect_equal(nrow(wide), 2)
     expect_equal(nrow(wide), nrow(wide_se))